backend/requirements.txt
fastapi==0.114.0
uvicorn[standard]==0.30.0
python-dotenv==1.0.1
stripe==10.10.0
python-multipart
certifi